refactor: move the experiment file to the main folder
