+++ Assets/Adefagia/Code/Scripts/RobotSystem/RobotController.cs
@@ -1,8 +1,6 @@
-﻿using System;
-using Adefagia.BattleMechanism;
-using Adefagia.Experimentation.dimaspadma.Agent;
+﻿using Adefagia.BattleMechanism;
+using Adefagia.PlayerAction;
 using Adefagia.GridSystem;
-using Unity.VisualScripting;
 using UnityEngine;
 using Grid = Adefagia.GridSystem.Grid;
